Refactor Result classes into result.py
